moved team and about page
