+++ pages/about.tsx
@@ -1,9 +1,4 @@
-import Image from "next/image";
 import Head from "next/head";
-
-import officers, { PersonInfoProps } from "@/data/officers";
-
-import styles from "@/styles/About.module.scss";
 
 type DataProps = {
   data: Record<string, any>;
@@ -25,41 +20,6 @@ export default function About({ data }: DataProps) {
         entirely different. We aim to nurture the love of cybersecurity in the
         students at UCLA.
       </p>
-
-      <h2>Team</h2>
-
-      <div className={styles.officersContainer}>
-        {officers.map((officer: PersonInfoProps, index: number) => (
-          <PersonInfo
-            key={index}
-            name={officer.name}
-            role={officer.role}
-            major={officer.major}
-            pronouns={officer.pronouns}
-            photo={officer.photo}
-          />
-        ))}
-      </div>
-    </div>
-  );
-}
-
-function PersonInfo({ name, role, major, pronouns, photo }: PersonInfoProps) {
-  return (
-    <div className={styles.personInfo}>
-      <Image
-        className={styles.personImage}
-        src={"/images/officers/" + photo}
-        alt={`Profile picture of ${name}`}
-        width={300}
-        height={300}
-      />
-      <h3>{name}</h3>
-      <p>
-        <i>{role}</i>
-      </p>
-      <p>{pronouns}</p>
-      <p>{major}</p>
     </div>
   );
 }
